refactor query conditions
query conditions are a subset of scan conditions
extract common conditions into a trait and then have both QueryCondition and ScanCondition objects
minor amount of renaming
